Move DataUpdateCoordinator to separate module in senz
